Moved util classes to util package
