+++ Warbots2020/src/main/java/frc/robot/bling/Bling.java
@@ -9,8 +9,8 @@ package frc.robot.bling;
 
 import edu.wpi.first.wpilibj.Spark;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.robot.Constants;
-import frc.robot.robot.Pin;
+import util.Constants;
+import util.Pin;
 
 public class Bling extends SubsystemBase 
 {
