+++ Warbots2020/src/main/java/frc/robot/commands/StuffFlyWheel.java
@@ -10,8 +10,8 @@ package frc.robot.commands;
 import java.time.LocalDateTime;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.robot.Constants;
 import frc.robot.subsystems.FlyWheel;
+import util.Constants;
 
 public class StuffFlyWheel extends CommandBase 
 {
