+++ Warbots2020/src/main/java/frc/robot/subsystems/FlyWheel.java
@@ -9,6 +9,9 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.robot.*;
+import util.Constants;
+import util.Pin;
+
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.*;
