+++ Warbots2020/src/main/java/frc/robot/subsystems/Intake.java
@@ -11,7 +11,7 @@ import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.robot.Pin;
+import util.Pin;
 
 public class Intake extends SubsystemBase
 {
