+++ Warbots2020/src/main/java/frc/robot/subsystems/Shooter.java
@@ -13,6 +13,8 @@ import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.robot.*;
+import util.Constants;
+import util.Pin;
 
 public class Shooter extends SubsystemBase 
 {
