+++ Warbots2020/src/main/java/util/Pin.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.robot;
+package util;
 
 import java.security.InvalidParameterException;
